Move classes to designated packages
