+++ src/main/java/com/ffxivcensus/gatherer/player/items/repositories/GearItemRepository.java
@@ -1,5 +1,6 @@
-package com.ffxivcensus.gatherer.player;
+package com.ffxivcensus.gatherer.player.items.repositories;
 
+import com.ffxivcensus.gatherer.player.items.data.GearItem;
 import org.springframework.data.repository.CrudRepository;
 
 public interface GearItemRepository extends CrudRepository<GearItem, String> {
